Change emit from callback to helpers file function
